Transform vhdre into module and package.
