+++ codegen/vhdre/__init__.py
@@ -1,0 +1,3 @@
+from .vhdre import RegexMatcher
+
+
